Fix test and clean up helper methods
I don't think we need to reset the authorize_called and policy_scoped
stack variables. Once the request completes the context will be popped
off anyway. On the other hand if there are more authorize calls in the
same request they should all 'pass' the test imposed by this method as
authorize was called at least once already. Note that the intention is
to check that authorize was called. If there is an exception raised by
the internal policy class method, that should not matter and hence the
stack context variable is set prior to actually delegating the call.