Allow for overriding automatic policy lookup with the __policy_class__ attribute
This will be a static property on a model class whose value should be
the policy class itself.
The real catch with in importing the policy class inside a model module
is also importing the model inside the policy class module and having a
cylic import error.
However, this really should not happen unless it is accidental. This is
because of the 'record' and 'scope' objects. These should be the only
references used inside the policy class. If there is any need to import
another model from a policy class, that should work fine as the cyclic
import is avoided (this action would be a bit strange though as 1 policy
should only ever have to deal with 1 model).